+++ src/main/java/qna/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package qna;
+package qna.exception;
 
 public class ForbiddenException extends RuntimeException {
     public ForbiddenException() {
